Optimizing code

+++ cache/cache.go
@@ -3,6 +3,8 @@ package cache
 import (
 	"context"
 	"time"
+
+	"github.com/dobyte/due/v2/errors"
 )
 
 var globalCache Cache
@@ -32,6 +34,8 @@ type Cache interface {
 	AddPrefix(key string) string
 	// Client 获取客户端
 	Client() any
+	// Close 关闭缓存
+	Close() error
 }
 
 // SetCache 设置缓存
@@ -46,55 +50,108 @@ func GetCache() Cache {
 
 // Has 检测缓存是否存在
 func Has(ctx context.Context, key string) (bool, error) {
+	if globalCache == nil {
+		return false, errors.ErrMissingCacheInstance
+	}
+
 	return globalCache.Has(ctx, key)
 }
 
 // Get 获取缓存值
 func Get(ctx context.Context, key string, def ...any) Result {
+	if globalCache == nil {
+		return NewResult(nil, errors.ErrMissingCacheInstance)
+	}
+
 	return globalCache.Get(ctx, key, def...)
 }
 
 // Set 设置缓存值
 func Set(ctx context.Context, key string, value any, expiration ...time.Duration) error {
+	if globalCache == nil {
+		return errors.ErrMissingCacheInstance
+	}
+
 	return globalCache.Set(ctx, key, value, expiration...)
 }
 
 // GetSet 获取设置缓存值
 func GetSet(ctx context.Context, key string, fn SetValueFunc) Result {
+	if globalCache == nil {
+		return NewResult(nil, errors.ErrMissingCacheInstance)
+	}
+
 	return globalCache.GetSet(ctx, key, fn)
 }
 
 // Delete 删除缓存
 func Delete(ctx context.Context, keys ...string) (int64, error) {
+	if globalCache == nil {
+		return 0, errors.ErrMissingCacheInstance
+	}
+
 	return globalCache.Delete(ctx, keys...)
 }
 
 // IncrInt 整数自增
 func IncrInt(ctx context.Context, key string, value int64) (int64, error) {
+	if globalCache == nil {
+		return 0, errors.ErrMissingCacheInstance
+	}
+
 	return globalCache.IncrInt(ctx, key, value)
 }
 
 // IncrFloat 浮点数自增
 func IncrFloat(ctx context.Context, key string, value float64) (float64, error) {
+	if globalCache == nil {
+		return 0, errors.ErrMissingCacheInstance
+	}
+
 	return globalCache.IncrFloat(ctx, key, value)
 }
 
 // DecrInt 整数自减
 func DecrInt(ctx context.Context, key string, value int64) (int64, error) {
+	if globalCache == nil {
+		return 0, errors.ErrMissingCacheInstance
+	}
+
 	return globalCache.DecrInt(ctx, key, value)
 }
 
 // DecrFloat 浮点数自减
 func DecrFloat(ctx context.Context, key string, value float64) (float64, error) {
+	if globalCache == nil {
+		return 0, errors.ErrMissingCacheInstance
+	}
+
 	return globalCache.DecrFloat(ctx, key, value)
 }
 
 // AddPrefix 添加Key前缀
 func AddPrefix(key string) string {
+	if globalCache == nil {
+		return ""
+	}
+
 	return globalCache.AddPrefix(key)
 }
 
 // Client 获取客户端
 func Client() any {
+	if globalCache == nil {
+		return nil
+	}
+
 	return globalCache.Client()
+}
+
+// Close 关闭缓存
+func Close() error {
+	if globalCache == nil {
+		return nil
+	}
+
+	return globalCache.Close()
 }
